requirements.txt
pre-commit
autoflake
black
flake8
Django
graphene
graphene-django
psycopg2
djangorestframework
django-cacheops
django-ckeditor
django-colorfield
django-extensions
django-graphql-jwt
django-jet
django-js-asset
django-modeladmin-reorder
django-nested-admin
django-crontab
django-graphql-jwt
python-dotenv
django-filter
